Refactor private/replace-contract tests in a separate file
